requirements.txt
agno
streamlit
yfinance
pandas
lxml
duckduckgo-search
google-genai
python-dotenv
streamlit
yfinance
plotly
requests
beautifulsoup4
html5lib
groq
numpy
openai
linkup-sdk

firecrawl-py
